Remove usage of xlog in SDK
