+++ exporters/stdout/stdoutlog/go.mod
@@ -34,5 +34,3 @@ replace go.opentelemetry.io/otel/trace => ../../../trace
 replace go.opentelemetry.io/otel/sdk => ../../../sdk
 
 replace go.opentelemetry.io/otel/metric => ../../../metric
-
-replace go.opentelemetry.io/otel/sdk/log/xlog => ../../../sdk/log/xlog
